+++ js/data.js
@@ -1,0 +1,42 @@
+const getData = () => {
+  const DECRIPTION_TEMPLATE = [
+    'Никогда не сдавайся, ведь ты не квартира.',
+    'Работа не волк. Волк - это гулять.',
+    'Одна ошибка и ты ошибся.',
+    'Уот так уот.',
+    'Если ты бездомный - просто купи дом.',
+    'Спаси и so horny.',
+    'Так исторически сложилось.',
+    'Never gonna give you up.',
+    'Sweet liberty, my arm!',
+    'Ктулху фхтагн.'
+  ];
+
+  const MESSAGE_TEMPLATE = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ];
+
+  const NAMES = [
+    'Вася',
+    'Игорь',
+    'Петя',
+    'Олег',
+    'Евгений',
+    'Кирилл',
+    'Тимур',
+    'Дима',
+    'Алексей',
+    'Виктор',
+    'Александр',
+    'Марсель',
+  ];
+
+  return {DECRIPTION_TEMPLATE, MESSAGE_TEMPLATE, NAMES};
+};
+
+export {getData};
